[#76] refactor: machines requestDto 디렉토리 이동
