+++ src/main/java/com/example/temp/machine/dto/request/MachineBulkCreateRequest.java
@@ -1,4 +1,4 @@
-package com.example.temp.admin.dto.request;
+package com.example.temp.machine.dto.request;
 
 import java.util.List;
