+++ src/test/java/com/example/temp/admin/dto/request/MachineCreateRequestTest.java
@@ -3,6 +3,7 @@ package com.example.temp.admin.dto.request;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.example.temp.machine.domain.BodyPart;
+import com.example.temp.machine.dto.request.MachineCreateRequest;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
