+++ src/test/java/com/example/temp/machine/application/MachineServiceTest.java
@@ -3,8 +3,8 @@ package com.example.temp.machine.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.example.temp.admin.dto.request.MachineBulkCreateRequest;
-import com.example.temp.admin.dto.request.MachineCreateRequest;
+import com.example.temp.machine.dto.request.MachineBulkCreateRequest;
+import com.example.temp.machine.dto.request.MachineCreateRequest;
 import com.example.temp.common.exception.ApiException;
 import com.example.temp.common.exception.ErrorCode;
 import com.example.temp.machine.domain.BodyPart;
